+++ lumicks/pylake/correlated_stack.py
@@ -1,7 +1,12 @@
 import numpy as np
 import os
 import re
+import json
+import cv2
+import tifffile
 import warnings
+
+from .channel import Slice, TimeSeries
 
 
 class TiffFrame:
@@ -13,12 +18,70 @@ class TiffFrame:
     page : tifffile.tifffile.TiffPage
         Tiff page recorded from a camera in Bluelake.
     """
-    def __init__(self, page):
+    def __init__(self, page, align):
         self._src = page
+        self._align = align
+
+    def _align_image(self):
+        """ reconstruct image using alignment matrices from Bluelake 
+            return aligned image as a NumPy array 
+        """
+
+        def correct_alignment_offset(alignment, x_offset, y_offset):
+            # translate the origin of the image so that it matches that of the original transform
+            translation = np.eye(3)
+            translation[0, -1] = -x_offset
+            translation[1, -1] = -y_offset
+            # apply the original transform to the translated image. 
+            # it only needs to be resized from a 2x3 to a 3x3 matrix
+            original = np.vstack((alignment, [0, 0, 1]))
+            # translate the image back to the original origin. 
+            # takes into account both the offset and the scaling performed by the first step
+            back_translation = np.eye(3)
+            back_translation[0, -1] = original[0, 0] * x_offset
+            back_translation[1, -1] = original[1, 1] * y_offset
+            # concatenate the transforms by multiplying their matrices and ignore the unused 3rd row
+            return np.dot(back_translation, np.dot(original, translation))[:2, :]
+
+        try:
+            description = json.loads(self._src.description)
+        except json.decoder.JSONDecodeError:
+            warnings.warn("File does not contain metadata. Only raw data is available")
+            return self.raw_data
+
+        try:
+            align_mats = [np.array(description[f"Alignment {color} channel"]).reshape((2,3))
+                        for color in ('red', 'blue')]
+            align_roi = np.array(description["Alignment region of interest (x, y, width, height)"])[:2]
+            roi = np.array(description["Region of interest (x, y, width, height)"])[:2]
+        except KeyError:
+            warnings.warn("File does not contain alignment matrices. Only raw data is available")
+            return self.raw_data
+        
+        x_offset, y_offset = align_roi - roi
+        if not (x_offset == 0 and y_offset == 0):
+            align_mats = [correct_alignment_offset(mat, x_offset, y_offset) for mat in align_mats]
+
+        img = self.raw_data
+        rows, cols, _ = img.shape
+        for mat, channel in zip(align_mats, (0, 2)):
+            img[:,:,channel] = cv2.warpAffine(img[:,:,channel], mat, (cols,rows),
+                                              flags=(cv2.INTER_LINEAR | cv2.WARP_INVERSE_MAP),
+                                              borderMode=cv2.BORDER_CONSTANT, borderValue=0)
+        return img
+        
 
     @property
     def data(self):
+        return self._align_image() if (self.is_rgb and self._align) else self._src.asarray()
+
+    @property
+    def raw_data(self):
         return self._src.asarray()
+
+    @property
+    def is_rgb(self):
+        return self._src.tags["SamplesPerPixel"].value == 3
 
     @property
     def start(self):
@@ -39,16 +102,16 @@ class TiffStack:
     tiff_file : tifffile.TiffFile
         TIFF file recorded from a camera in Bluelake.
     """
-    def __init__(self, tiff_file):
+    def __init__(self, tiff_file, align):
         self._src = tiff_file
+        self._align = align
 
     def get_frame(self, frame):
-        return TiffFrame(self._src.pages[frame])
+        return TiffFrame(self._src.pages[frame], align=self._align)
 
     @staticmethod
-    def from_file(image_file):
-        import tifffile
-        return TiffStack(tifffile.TiffFile(image_file))
+    def from_file(image_file, align):
+        return TiffStack(tifffile.TiffFile(image_file), align=align)
 
     @property
     def num_frames(self):
@@ -63,6 +126,10 @@ class CorrelatedStack:
     ----------
     image_name : str
         Filename for the image stack. Typically a TIFF file recorded from a camera in Bluelake.
+
+    align : bool
+        If enabled, multi-channel images will be reconstructed from the image alignment metadata
+        from Bluelake. The default value is `True`.
 
     Examples
     --------
@@ -80,8 +147,8 @@ class CorrelatedStack:
         # Determine the force trace averaged over frame 2...9.
         file.force1x.downsampled_over(stack[2:10].timestamps)
     """
-    def __init__(self, image_name):
-        self.src = TiffStack.from_file(image_name)
+    def __init__(self, image_name, align=True):
+        self.src = TiffStack.from_file(image_name, align=align)
         self.name = os.path.splitext(os.path.basename(image_name))[0]
         self.start_idx = 0
         self.stop_idx = self.src.num_frames